requirements.txt
beautifulsoup4>=4.12.0
Flask>=2.0.0
flask-cors>=3.0.10
lxml>=4.9.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0